requirements.txt
pycryptodome
customtkinter
bcrypt
